# Optimize add_items for Varcher by automatically avoiding deep copying.