Summary:
Before:
```
array_concat_VARCHAR_10##2arg                             652.44ms      1.53
array_concat_VARCHAR_10##3arg                             958.59ms      1.04
array_concat_VARCHAR_10##4arg                                1.26s   790.86m
array_concat_VARCHAR_20##2arg                                1.67s   598.25m
array_concat_VARCHAR_20##3arg                                2.22s   449.48m
array_concat_VARCHAR_20##4arg                                2.82s   355.01m
array_concat_VARCHAR_40##2arg                                2.83s   353.24m
array_concat_VARCHAR_40##3arg                                4.98s   200.99m
array_concat_VARCHAR_40##4arg                                7.03s   142.22m
array_concat_VARCHAR_5##2arg                              290.04ms      3.45
array_concat_VARCHAR_5##3arg                              438.06ms      2.28
array_concat_VARCHAR_5##4arg                              584.20ms      1.71
````

After:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
array_concat_VARCHAR_10##2arg                             228.69ms      4.37
array_concat_VARCHAR_10##3arg                             305.58ms      3.27
array_concat_VARCHAR_10##4arg                             392.84ms      2.55
array_concat_VARCHAR_20##2arg                             580.45ms      1.72
array_concat_VARCHAR_20##3arg                             749.82ms      1.33
array_concat_VARCHAR_20##4arg                                1.00s   999.14m
array_concat_VARCHAR_40##2arg                                1.15s   867.18m
array_concat_VARCHAR_40##3arg                                1.83s   545.75m
array_concat_VARCHAR_40##4arg                                2.15s   464.43m
array_concat_VARCHAR_5##2arg                              101.89ms      9.81
array_concat_VARCHAR_5##3arg                              139.85ms      7.15
array_concat_VARCHAR_5##4arg                              185.80ms      5.38
```

Differential Revision: D50957600